Revert "refactor: 合併 StatisticsQueryService 至 StatisticsApplicationService (#125)"
This reverts commit 3d2fd5796d8f51741690ff4a43f9fd99e5f15314.
