+++ backend/app/Infrastructure/Statistics/Adapters/StatisticsQueryAdapter.php
@@ -4,16 +4,16 @@ declare(strict_types=1);
 
 namespace App\Infrastructure\Statistics\Adapters;
 
-use App\Application\Services\Statistics\DTOs\StatisticsQueryDTO;
-use App\Application\Services\Statistics\StatisticsApplicationService;
+use App\Domains\Statistics\Contracts\StatisticsQueryServiceInterface as BaseQueryServiceInterface;
 use DateInterval;
+use DateTime;
 use DateTimeImmutable;
 use DateTimeInterface;
 
 class StatisticsQueryAdapter
 {
     public function __construct(
-        private StatisticsApplicationService $statisticsApplicationService,
+        private BaseQueryServiceInterface $baseQueryService,
     ) {}
 
     /**
@@ -26,12 +26,13 @@ class StatisticsQueryAdapter
         DateTimeInterface $endDate,
         string $granularity,
     ): array {
-        $queryDTO = new StatisticsQueryDTO(
-            startDate: DateTimeImmutable::createFromInterface($startDate),
-            endDate: DateTimeImmutable::createFromInterface($endDate),
-        );
-        $this->statisticsApplicationService->getPostStatistics($queryDTO);
+        $options = [
+            'period_start' => DateTime::createFromInterface($startDate),
+            'period_end'   => DateTime::createFromInterface($endDate),
+        ];
+        $data = $this->baseQueryService->getPostStatistics($options);
 
+        // 模擬時間序列資料生成
         return $this->generateMockTimeSeriesData($startDate, $endDate, $granularity, 'posts');
     }
 
@@ -45,11 +46,11 @@ class StatisticsQueryAdapter
         DateTimeInterface $endDate,
         string $granularity,
     ): array {
-        $queryDTO = new StatisticsQueryDTO(
-            startDate: DateTimeImmutable::createFromInterface($startDate),
-            endDate: DateTimeImmutable::createFromInterface($endDate),
-        );
-        $this->statisticsApplicationService->getUserStatistics($queryDTO);
+        $options = [
+            'period_start' => DateTime::createFromInterface($startDate),
+            'period_end'   => DateTime::createFromInterface($endDate),
+        ];
+        $data = $this->baseQueryService->getUserStatistics($options);
 
         return $this->generateMockTimeSeriesData($startDate, $endDate, $granularity, 'users');
     }
@@ -64,11 +65,14 @@ class StatisticsQueryAdapter
         ?DateTimeInterface $endDate = null,
         int $limit = 10,
     ): array {
-        $queryDTO = new StatisticsQueryDTO(
-            startDate: $startDate !== null ? DateTimeImmutable::createFromInterface($startDate) : null,
-            endDate: $endDate !== null ? DateTimeImmutable::createFromInterface($endDate) : null,
-        );
-        $this->statisticsApplicationService->getSourceDistribution($queryDTO);
+        $options = [];
+        if ($startDate) {
+            $options['period_start'] = DateTime::createFromInterface($startDate);
+        }
+        if ($endDate) {
+            $options['period_end'] = DateTime::createFromInterface($endDate);
+        }
+        $data = $this->baseQueryService->getSourceDistribution($options);
 
         return $this->generateMockCategoryData('source', $limit);
     }
@@ -123,11 +127,14 @@ class StatisticsQueryAdapter
         string $sortBy = 'views',
         int $limit = 10,
     ): array {
-        $queryDTO = new StatisticsQueryDTO(
-            startDate: $startDate !== null ? DateTimeImmutable::createFromInterface($startDate) : null,
-            endDate: $endDate !== null ? DateTimeImmutable::createFromInterface($endDate) : null,
-        );
-        $this->statisticsApplicationService->getPopularContent($queryDTO);
+        $options = [];
+        if ($startDate) {
+            $options['period_start'] = DateTime::createFromInterface($startDate);
+        }
+        if ($endDate) {
+            $options['period_end'] = DateTime::createFromInterface($endDate);
+        }
+        $data = $this->baseQueryService->getPopularContent($options);
 
         return $this->generateMockCategoryData('content', $limit);
     }
