--- openspec/changes/merge-statistics-query-service/.openspec.yaml
@@ -1,2 +1,0 @@
-schema: spec-driven
-created: 2026-07-16
